requirements.txt
numpy==1.22.4
matplotlib==3.7.1
seaborn==0.12.2
torch==2.0.0
torchvision==0.15.1
tqdm
scikit-learn==1.2.2
Pillow # Note: "PIL" is actually accessed through the "Pillow" package on PyPI.
opencv-python==4.10.0.84
flask>=3.1.0